requirements/base.txt
uvicorn
fastapi
python-multipart
python-dotenv
sqlalchemy
python-jose
passlib
bcrypt
iso3166
jinja2
pydantic[email]
pandas
openpyxl